Benchmark CREPS inherit from optimizer CREPS
